+++ lib/presentation/screen/setting_screen.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:star_book/presentation/shared/tile.dart';
+import 'package:star_book/widgets/gradient_scaffold.dart';
 
 class SettingScreen extends StatelessWidget {
   const SettingScreen({Key? key}) : super(key: key);
@@ -6,8 +8,7 @@ class SettingScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final double screenHeight = MediaQuery.of(context).size.height;
-    return Scaffold(
-      backgroundColor: Colors.transparent,
+    return GradientScaffold(
       appBar: AppBar(
         elevation: 0,
         backgroundColor: Colors.transparent,
@@ -54,134 +55,6 @@ class SettingScreen extends StatelessWidget {
           const SizedBox(height: 10),
           const UxerShipTile(),
         ],
-      ),
-    );
-  }
-}
-
-class BlocShipTile extends StatelessWidget {
-  const BlocShipTile({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        const Image(
-          image: AssetImage('assets/icons/blocship_logo.png'),
-          width: 16,
-        ),
-        const SizedBox(width: 6),
-        RichText(
-          textScaleFactor: MediaQuery.of(context).textScaleFactor,
-          text: TextSpan(
-            text: 'Powered by ',
-            style: Theme.of(context)
-                .textTheme
-                .bodyMedium!
-                .copyWith(fontWeight: FontWeight.w400),
-            children: [
-              TextSpan(
-                text: 'Blocship',
-                style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-                      fontWeight: FontWeight.w400,
-                      color: const Color(0xFF1F1F1F),
-                      decoration: TextDecoration.underline,
-                    ),
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class UxerShipTile extends StatelessWidget {
-  const UxerShipTile({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        const Image(
-          image: AssetImage('assets/icons/ux_logo.png'),
-          width: 22,
-        ),
-        const SizedBox(width: 6),
-        RichText(
-          textScaleFactor: MediaQuery.of(context).textScaleFactor,
-          text: TextSpan(
-            text: 'Design & Crafted with ',
-            style: Theme.of(context)
-                .textTheme
-                .bodyMedium!
-                .copyWith(fontWeight: FontWeight.w400),
-            children: [
-              const WidgetSpan(
-                child: Icon(
-                  Icons.favorite,
-                  color: Colors.red,
-                  size: 17,
-                ),
-              ),
-              TextSpan(
-                  text: ' by ',
-                  style: Theme.of(context)
-                      .textTheme
-                      .bodyMedium!
-                      .copyWith(fontWeight: FontWeight.w400)),
-              TextSpan(
-                text: 'Uxership',
-                style: Theme.of(context).textTheme.bodyMedium!.copyWith(
-                      fontWeight: FontWeight.w400,
-                      color: const Color(0xFF1F1F1F),
-                      decoration: TextDecoration.underline,
-                    ),
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class CustomTile extends StatelessWidget {
-  final String title;
-  final String subtitle;
-  final VoidCallback onTap;
-
-  const CustomTile({
-    Key? key,
-    required this.title,
-    required this.subtitle,
-    required this.onTap,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: MediaQuery.of(context).size.width * 0.95,
-      child: ListTile(
-        title: Text(
-          title,
-          style: Theme.of(context).textTheme.bodyLarge!.copyWith(
-              fontWeight: FontWeight.w600, color: const Color(0xFF1F1F1F)),
-        ),
-        subtitle: Text(
-          subtitle,
-          style: Theme.of(context)
-              .textTheme
-              .bodyMedium!
-              .copyWith(fontWeight: FontWeight.w400),
-        ),
-        trailing: const Icon(
-          Icons.keyboard_arrow_right,
-          color: Color(0xFF1F1F1F),
-        ),
-        onTap: () => onTap,
       ),
     );
   }
